new update

+++ app/Http/Controllers/Profile/ProfileController.php
@@ -3,11 +3,11 @@
 namespace App\Http\Controllers\Profile;
 
 use App\Http\Controllers\Controller;
-use App\Models\UserModel;
 use App\Repository\UserRepository;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class ProfileController extends Controller
 {
@@ -26,23 +26,34 @@ class ProfileController extends Controller
         return view('profile.profile');
     }
 
-    public function update()
+    public function update($type)
     {
         $data = [];
         $data = $this->request->all();
-        $data['id'] = Auth::user()->id;
-        if ($this->request->hasFile('ft_user')) {
-            $file = $this->request->file('ft_user');
-            $fileName = $data['nm_user'].'.'.$file->getClientOriginalExtension();
-            if (Storage::exists('public/img/foto'. $fileName)) {
-                Storage::delete('public/img/foto/'. $fileName);
+        $id = Auth::user()->id;
+        $validator = Validator::make($data, [
+            'password'=>'required',
+            'passwordretype'=>'required|same:password'
+        ]);
+        if ($type == 'profile') {
+            if ($this->request->hasFile('ft_user')) {
+                $file = $this->request->file('ft_user');
+                $fileName = $data['nm_user'].'.'.$file->getClientOriginalExtension();
+                if (Storage::exists('public/img/foto'. $fileName)) {
+                    Storage::delete('public/img/foto/'. $fileName);
+                }
+                $file->storeAs('public/img/foto/', $fileName);
+                $data['ft_user'] = $fileName;
+                $this->userRepo->Update($data, $id);
+                return response()->json(['success'=> ['message'=>'Profile Berhasil Di Perbaharui']], 200);
             }
-            $file->storeAs('public/img/foto/', $fileName);
-            $data['ft_user'] = $fileName;
-            $this->userRepo->Update($data);
-            return response()->json(['success'=> ['message'=>'Profile Berhasil Di Perbaharui']], 200);
+            $this->userRepo->update($data, $id);
+            return response()->json(['success'=>['message'=>'Profile Berhasil Di Perbaharui']], 200);
+        }else{
+            if ($validator->fails()) {
+                return response()->json(['errors'=>$validator->errors()], 422);
+            }
+            // dd($data);
         }
-        $this->userRepo->update($data);
-        return response()->json(['success'=>['message'=>'Profile Berhasil Di Perbaharui']], 200);
     }
 }
